refactor:Seperated response functions from controller to ApiTraits.fix:error handler added for RouteNotFound Exception
